+++ de4a-connector/src/main/java/eu/de4a/connector/api/manager/EvidenceTransferorManager.java
@@ -11,16 +11,16 @@ import org.springframework.stereotype.Component;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 
-import eu.de4a.connector.api.controller.error.ErrorHandlerUtils;
-import eu.de4a.connector.api.controller.error.ExternalModuleError;
-import eu.de4a.connector.api.controller.error.FamilyErrorType;
-import eu.de4a.connector.api.controller.error.ResponseErrorException;
-import eu.de4a.connector.api.controller.error.ResponseTransferEvidenceException;
 import eu.de4a.connector.as4.client.regrep.RegRepTransformer;
 import eu.de4a.connector.as4.owner.MessageOwner;
 import eu.de4a.connector.as4.owner.MessageResponseOwner;
 import eu.de4a.connector.as4.owner.OwnerLocator;
 import eu.de4a.connector.client.Client;
+import eu.de4a.connector.error.exceptions.ResponseErrorException;
+import eu.de4a.connector.error.exceptions.ResponseTransferEvidenceException;
+import eu.de4a.connector.error.model.ExternalModuleError;
+import eu.de4a.connector.error.model.FamilyErrorType;
+import eu.de4a.connector.error.utils.ErrorHandlerUtils;
 import eu.de4a.connector.model.OwnerAddresses;
 import eu.de4a.connector.model.RequestorRequest;
 import eu.de4a.connector.model.smp.NodeInfo;
