+++ de4a-connector/src/main/java/eu/de4a/connector/as4/client/regrep/RegRepTransformer.java
@@ -25,10 +25,10 @@ import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.xml.sax.InputSource;
 
-import eu.de4a.connector.api.controller.error.ConnectorException;
-import eu.de4a.connector.api.controller.error.ExternalModuleError;
-import eu.de4a.connector.api.controller.error.FamilyErrorType;
-import eu.de4a.connector.api.controller.error.LayerError;
+import eu.de4a.connector.error.exceptions.ConnectorException;
+import eu.de4a.connector.error.model.ExternalModuleError;
+import eu.de4a.connector.error.model.FamilyErrorType;
+import eu.de4a.connector.error.model.LayerError;
 import eu.de4a.exception.MessageException;
 import eu.de4a.util.DOMUtils;
