+++ de4a-connector/src/main/java/eu/de4a/connector/client/Client.java
@@ -41,15 +41,15 @@ import com.helger.smpclient.url.SMPDNSResolutionException;
 import com.helger.xsds.bdxr.smp1.EndpointType;
 import com.helger.xsds.bdxr.smp1.SignedServiceMetadataType;
 
-import eu.de4a.connector.api.controller.error.ConnectorException;
-import eu.de4a.connector.api.controller.error.ErrorHandlerUtils;
-import eu.de4a.connector.api.controller.error.ExternalModuleError;
-import eu.de4a.connector.api.controller.error.FamilyErrorType;
-import eu.de4a.connector.api.controller.error.LayerError;
-import eu.de4a.connector.api.controller.error.ResponseErrorException;
-import eu.de4a.connector.api.controller.error.ResponseExtractEvidenceException;
-import eu.de4a.connector.api.controller.error.ResponseLookupRoutingInformationException;
-import eu.de4a.connector.api.controller.error.SMPLookingMetadataInformationException;
+import eu.de4a.connector.error.exceptions.ConnectorException;
+import eu.de4a.connector.error.exceptions.ResponseErrorException;
+import eu.de4a.connector.error.exceptions.ResponseExtractEvidenceException;
+import eu.de4a.connector.error.exceptions.ResponseLookupRoutingInformationException;
+import eu.de4a.connector.error.exceptions.SMPLookingMetadataInformationException;
+import eu.de4a.connector.error.model.ExternalModuleError;
+import eu.de4a.connector.error.model.FamilyErrorType;
+import eu.de4a.connector.error.model.LayerError;
+import eu.de4a.connector.error.utils.ErrorHandlerUtils;
 import eu.de4a.connector.model.smp.NodeInfo;
 import eu.de4a.iem.jaxb.common.types.AckType;
 import eu.de4a.iem.jaxb.common.types.AvailableSourcesType;
