+++ de4a-connector/src/main/java/eu/de4a/connector/error/exceptions/ResponseTransferEvidenceException.java
@@ -1,4 +1,4 @@
-package eu.de4a.connector.api.controller.error;
+package eu.de4a.connector.error.exceptions;
 
 public class ResponseTransferEvidenceException extends ConnectorException {  
 	private static final long serialVersionUID = 1L;
