+++ de4a-connector/src/main/java/eu/de4a/connector/error/handler/ResponseErrorExceptionHandler.java
@@ -1,5 +1,6 @@
-package eu.de4a.connector.api.controller.error;
+package eu.de4a.connector.error.handler;
 
+import eu.de4a.connector.error.exceptions.ConnectorException;
 import eu.de4a.iem.jaxb.common.types.ErrorListType;
 import eu.de4a.iem.jaxb.common.types.ResponseErrorType;
 import eu.de4a.iem.xml.de4a.DE4AMarshaller;
