+++ de4a-connector/src/main/java/eu/de4a/connector/error/handler/ResponseTransferEvidenceExceptionHandler.java
@@ -1,5 +1,6 @@
-package eu.de4a.connector.api.controller.error;
+package eu.de4a.connector.error.handler;
 
+import eu.de4a.connector.error.exceptions.ConnectorException;
 import eu.de4a.iem.jaxb.common.types.ErrorListType;
 import eu.de4a.iem.jaxb.common.types.RequestTransferEvidenceUSIIMDRType;
 import eu.de4a.iem.jaxb.common.types.ResponseTransferEvidenceType;
