+++ de4a-connector/src/main/java/eu/de4a/connector/error/model/FamilyErrorType.java
@@ -1,4 +1,4 @@
-package eu.de4a.connector.api.controller.error;
+package eu.de4a.connector.error.model;
 
 public enum FamilyErrorType { 
 	MISSING_REQUIRED_ARGUMENTS("2","error.missing.arguments"),
