+++ de4a-connector/src/main/java/eu/de4a/connector/error/utils/ResponseErrorFactory.java
@@ -1,7 +1,21 @@
-package eu.de4a.connector.api.controller.error;
+package eu.de4a.connector.error.utils;
 
 import java.util.HashMap;
 import java.util.Map;
+
+import eu.de4a.connector.error.exceptions.ConnectorException;
+import eu.de4a.connector.error.exceptions.ResponseErrorException;
+import eu.de4a.connector.error.exceptions.ResponseExtractEvidenceException;
+import eu.de4a.connector.error.exceptions.ResponseLookupRoutingInformationException;
+import eu.de4a.connector.error.exceptions.ResponseTransferEvidenceException;
+import eu.de4a.connector.error.exceptions.SMPLookingMetadataInformationException;
+import eu.de4a.connector.error.handler.ConnectorExceptionHandler;
+import eu.de4a.connector.error.handler.GenericExceptionHandler;
+import eu.de4a.connector.error.handler.ResponseErrorExceptionHandler;
+import eu.de4a.connector.error.handler.ResponseExtractEvidenceExceptionHandler;
+import eu.de4a.connector.error.handler.ResponseLookupRoutingInformationExceptionHandler;
+import eu.de4a.connector.error.handler.ResponseTransferEvidenceExceptionHandler;
+import eu.de4a.connector.error.handler.SMPLookingMetadataInformationExceptionHandler;
 
 public class ResponseErrorFactory {
 	private static Map<Class<?>,ConnectorExceptionHandler> handlers;
